[bus-updater] FIX loop() checks only telegramReadyToSend now
-added more serial logging
-first attempt of catch unit tests for BcuLayer4
-added TS_ARM support with #define TS_ARM
-fixed catch unit test for sblib not compiling, because of changed parameters of IAP_Call(.)
-added transfer speed output to pc_updater_tool
-fixed typos

+++ test/sblib/cpu-emu/system_lpc11xx.c
@@ -72,7 +72,7 @@ void IAP_Init_Flash(unsigned char value)
     memset(FLASH, value, FLASH_SIZE);
 }
 
-void IAP_Call (unsigned int * cmd, unsigned int * stat)
+void IAP_Call (unsigned long * cmd, unsigned long * stat)
 {
     unsigned int i;
     unsigned int end;
